pyproject.toml
[project]
name = "midjourneyactivityemulator"
version = "1.0"
description = "Emulates activity for midjourney"
authors = [
    {name = "github.com/UndoGG",email = "hey.ya.dj@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "click (>=8.1.8,<9.0.0)",
    "rich (>=13.9.4,<14.0.0)",
    "aiohttp (>=3.11.14,<4.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
